+++ src/entities.rs
@@ -4,5 +4,5 @@ use std::collections::HashMap;
 pub struct Collector {
     pub user_id: i64,
     pub chat_id: i64,
-    pub stickers: HashMap<String, Vec<u16>>
+    pub stickers: HashMap<String, Vec<u64>>
 }
